Implimented cookie functionality in the HttpRequest class + additional stfuff.
Now the class has a Cookies property. Cookies will be parsed from the HeaderCollection if there are any.
Added a HttpConstants class with RequestCookieHeader and NewLine properties.
HttpResponse class is now protected insted of public.

P.S.: Basically what I did is to get the cookies implementation of a previous SUS and I "parsed" it to this one.