Added a check for a password which consists of only whitespaces.
I Added a check for a password which consists of only whitespaces in the ValidateUser method.

I used a Linq expression because String.IsNullOrWhiteSpace() will return true if the password input is null which will add an incorect error -> "The provided password cannot be only whitespaces!".

Now the if statement in the HashPassword class is not needed and thus I deleted it.